Rename Project

+++ neo/swf/SWF_Inflate.cpp
@@ -40,7 +40,7 @@ bool idSWF::Inflate( const byte* input, int inputSize, byte* output, int outputS
 {
 	struct local_swf_alloc_t
 	{
-		static void* zalloc( void* opaque, uint32 items, uint32 size )
+		static void* zalloc( void* opaque, size_t items, size_t size )
 		{
 			return Mem_Alloc( items * size, TAG_SWF );
 		}
@@ -73,7 +73,7 @@ bool idSWF::Deflate( const byte* input, int inputSize, byte* output, int& output
 {
 	struct local_swf_alloc_t
 	{
-		static void* zalloc( void* opaque, uint32 items, uint32 size )
+		static void* zalloc( void* opaque, size_t items, size_t size )
 		{
 			return Mem_Alloc( items * size, TAG_SWF );
 		}
